Trait jsonrpsee_http_client::traits::SubscriptionClient

pub trait SubscriptionClient: Client {
#[must_use]    pub fn subscribe<'a, 'life0, 'async_trait, Notif>(
        &'life0 self, 
        subscribe_method: &'a str, 
        params: JsonRpcParams<'a>, 
        unsubscribe_method: &'a str
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<Notif>, Error>> + 'async_trait + Send, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        Notif: DeserializeOwned + 'async_trait;
}

JSON-RPC client interface that can make requests, notifications and subscriptions.

Required methods

#[must_use]pub fn subscribe<'a, 'life0, 'async_trait, Notif>(
&'life0 self,
subscribe_method: &'a str,
params: JsonRpcParams<'a>,
unsubscribe_method: &'a str
) -> Pin<Box<dyn Future<Output = Result<Subscription<Notif>, Error>> + 'async_trait + Send, Global>> where
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
Notif: DeserializeOwned + 'async_trait,

Send a subscription request to the server, technically not part of the JSON-RPC specification

The subscribe_method and params are used to ask for the subscription towards the server.

The unsubscribe_method is used to close the subscription.

The Notif param is a generic type to receive generic subscriptions, see Subscription for further documentation.

Loading content...