Struct jsonrpsee_http_client::transport::HttpTransportClientBuilder

pub struct HttpTransportClientBuilder<L> { /* private fields */ }

Builder for HttpTransportClient.

Implementations

impl HttpTransportClientBuilder<Identity>

pub fn new() -> Self

Create a new HttpTransportClientBuilder.

impl<L> HttpTransportClientBuilder<L>

pub fn set_certification_store(self, cert_store: CertificateStore) -> Self

Set the certificate store.

pub fn max_request_size(self, size: u32) -> Self

Set the maximum size of a request body in bytes. Default is 10 MiB.

pub fn max_response_size(self, size: u32) -> Self

Set the maximum size of a response in bytes. Default is 10 MiB.

pub fn set_headers(self, headers: HeaderMap) -> Self

Set a custom header passed to the server with every request (default is none).

The caller is responsible for checking that the headers do not conflict or are duplicated.

pub fn set_tcp_no_delay(self, no_delay: bool) -> Self

Configure TCP_NODELAY on the socket to the supplied value nodelay.

Default is true.

pub fn set_max_logging_length(self, max: u32) -> Self

Max length for logging for requests and responses in number characters.

Logs bigger than this limit will be truncated.

pub fn set_service<T>( self, service: ServiceBuilder<T> ) -> HttpTransportClientBuilder<T>

Configure a tower service.

pub fn build<S, B>( self, target: impl AsRef<str> ) -> Result<HttpTransportClient<S>, Error>
where L: Layer<HttpBackend<Body>, Service = S>, S: Service<Request<Body>, Response = Response<B>, Error = Error> + Clone, B: HttpBody + Send + 'static, B::Data: Send, B::Error: Into<Box<dyn StdError + Send + Sync>>,

Build a HttpTransportClient.