Crate jsonrpsee_http_client

jsonrpsee-http-client

jsonrpsee-http-client is JSON RPC HTTP client library that's is built for async/await.

It is tightly-coupled to tokio because hyper is used as transport client, which is not compatible with other async runtimes such as async-std, smol and similar.

It supports both tokio 1.0 and tokio 0.2 via Optional features.

Optional Features

jsonrpsee-http-client uses the following feature flags:

tokio1: Enable to use the library with tokio 1.0 (mutually exclusive with tokio02)

tokio0.2: Enable to use the library with tokio 0.2 (mutually exclusive with tokio1)

Modules

error
Shared error type.

traits
Traits

v2
JSON-RPC 2.0 specification related types v2.

Structs

BatchMessage
Batch request message.

HttpClient
JSON-RPC HTTP Client that provides functionality to perform method calls and notifications.

HttpClientBuilder
Http Client Builder.

JsonRawValue
Reference to a range of bytes encompassing a single valid JSON value in the input data.

RequestMessage
Request message.

Subscription
Active subscription on a Client.

SubscriptionMessage
Subscription message.

Enums

Error
Error type.

FrontToBack
Message that the Client can send to the background task.

JsonValue
Represents any valid JSON value.

Constants

TEN_MB_SIZE_BYTES
Ten megabytes.

Traits

DeserializeOwned
A data structure that can be deserialized without borrowing any data from the deserializer.

Serialize
A data structure that can be serialized into any data format supported by Serde.

Functions

to_json_value
Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

Type Definitions

Cow
Compact three word Cow that puts the ownership tag in capacity. This is a type alias, for documentation see beef::generic::Cow.

Derive Macros

Serialize