Crate jsonrpsee_http_client

jsonrpsee-http-client

jsonrpsee-http-client is JSON RPC HTTP client library that’s is built for async/await.

It is tightly-coupled to tokio because hyper is used as transport client, which is not compatible with other async runtimes such as async-std, smol and similar.

Re-exports

pub use jsonrpsee_types as types;

Modules

transport

HTTP transport.

Structs

HeaderMap

A set of HTTP headers

HeaderValue

Represents an HTTP header field value.

HttpClient

JSON-RPC HTTP Client that provides functionality to perform method calls and notifications.

HttpClientBuilder

HTTP client builder.